test(mfa): move validate-code loading test
